requirements.txt
# Core dependencies
youtube-transcript-api==0.6.2
click==8.1.7
rich==13.7.0
requests==2.31.0
python-dotenv==1.0.0

# LLM providers (optional)
anthropic==0.8.1
google-generativeai==0.3.2

# Development dependencies
pytest==7.4.3
pytest-mock==3.12.0
black==23.12.0
flake8==6.1.0
